Changed period in row evolution breaks graph/metric association
Started working on it but it's not fully working yet. While it keeps it in memory when switching between visualisations, the selected metrics and the shown metrics might be out of sync. When switching between periods, `this.jqplotParams.series` seems to contain different metrics. Eg 5 metrics for all non-day metrics and 7 metrics for all other metrics. As such when I remember the order of which ones are activated, things get out of sync.

Also the shown metrics below the graph aren't correct anymore and it would sometimes need to show less metrics by the looks. So we would possibly need to hide some metrics on demand. 

When saving which metrics are currently selected, we probably need to store metric names instead of indexes to correctly highlight the correct ones. This seems quite challenging though. 

Tried to fix these things but after 4 hours couldn't fix it just yet. I reckon this needs a lot more work. I wonder if we even need to reload the entire popover to make this work correctly. Might be easier then trying to emulate everything else.