Prevent warning array to string conversion in console renderer (#15037)
See https://forum.matomo.org/t/notice-array-to-string-conversion-matomo-3-11-0/34644

As we don't have more information yet I couldn't reproduce it. Looking at the code it seems though like some metadata has an array in there. I didn't even quite know we had this renderer and from what I can see we're using it to convert a dataTableMap to a string. Not sure when and why. It should fix the warning though.